+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamo/impl/collections/facet/model/FacetFactory.java
@@ -96,7 +96,7 @@ public final class FacetFactory {
 
 	private <D extends DtObject> DtList<D> apply(final ListFilter listFilter, final DtList<D> fullDtList) {
 		//on délégue à CollectionsManager les méthodes de requête de filtrage.
-		return collectionManager.createDtListProcessor()
+		return collectionManager.<D> createDtListProcessor()
 				.filter(listFilter)
 				.apply(fullDtList);
 	}